# internal/api/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "immerle.ActivityResponse": {
                "properties": {
                    "events": {
                        "items": {
                            "$ref": "#/components/schemas/immerle.ActivityEventDTO"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "ok": {
                        "example": true,
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "immerle.ErrorResponse": {
                "properties": {
                    "error": {
                        "example": "already_initialized",
                        "type": "string"
                    },
                    "ok": {
                        "example": false,
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "immerle.FriendDTO": {
                "properties": {
                    "displayName": {
                        "example": "Bob Marley",
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "username": {
                        "example": "bob",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "immerle.FriendsResponse": {
                "properties": {
                    "friends": {
                        "items": {
                            "$ref": "#/components/schemas/immerle.FriendDTO"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "ok": {
                        "example": true,
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "immerle.OKResponse": {
                "properties": {
                    "ok": {
                        "example": true,
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "immerle.PendingFriendDTO": {
                "properties": {
                    "displayName": {
                        "example": "Alice Coltrane",
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "since": {
                        "example": "2026-06-15T09:00:00Z",
                        "type": "string"
                    },
                    "username": {
                        "example": "alice",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "immerle.PendingFriendsResponse": {
                "properties": {
                    "ok": {
                        "example": true,
                        "type": "boolean"
                    },
                    "pending": {
                        "items": {
                            "$ref": "#/components/schemas/immerle.PendingFriendDTO"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "contact": {
            "name": "immerle"
        },
        "description": "Native immerle endpoints (first-run setup, capability\ndiscovery, friends, activity feed, collaborative playlists and\nsynchronized Jam sessions) that complement the Subsonic /\nOpenSubsonic API served under /rest/.\n\nAuthenticated endpoints reuse Subsonic credentials passed as\nquery parameters: u (username) plus either p (password) or the\ntoken pair t (md5(password+salt)) and s (salt), and c (client\nname). Setup and capability endpoints are unauthenticated.",
        "license": {
            "name": "See repository"
        },
        "title": "immerle extension API",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/activity": {
            "get": {
                "description": "Returns activity events visible to the caller, honoring each author's privacy setting.",
                "parameters": [
                    {
                        "description": "Subsonic username",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password (or t+s token auth)",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ActivityResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Activity feed",
                "tags": [
                    "activity"
                ]
            }
        },
...
        "/friends": {
            "get": {
                "description": "Returns the caller's accepted friends.",
                "parameters": [
                    {
                        "description": "Subsonic username",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password (or use t+s token auth)",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.FriendsResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "summary": "List friends",
                "tags": [
                    "friends"
                ]
            }
        },
        "/friends/accept": {
            "post": {
                "parameters": [
                    {
                        "description": "Subsonic username",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password (or t+s token auth)",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Username of the requester to accept",
                        "in": "query",
                        "name": "username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.OKResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Accept a friend request",
                "tags": [
                    "friends"
                ]
            }
        },
        "/friends/pending": {
            "get": {
                "parameters": [
                    {
                        "description": "Subsonic username",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password (or t+s token auth)",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.PendingFriendsResponse"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "List pending friend requests",
                "tags": [
                    "friends"
                ]
            }
        },
        "/friends/request": {
            "post": {
                "parameters": [
                    {
                        "description": "Subsonic username",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password (or t+s token auth)",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Target username to befriend",
                        "in": "query",
                        "name": "username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.OKResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Send a friend request",
                "tags": [
                    "friends"
                ]
            }
        },
...
        "/playlists/collaborators": {
            "post": {
                "description": "Owner-only. Marks the playlist collaborative and grants edit rights to another user.",
                "parameters": [
                    {
                        "description": "Subsonic username",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password (or t+s token auth)",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Playlist id",
                        "in": "query",
                        "name": "playlistId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "User to grant edit rights",
                        "in": "query",
                        "name": "username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.OKResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "Add a playlist collaborator",
                "tags": [
                    "playlists"
                ]
            }
        },
...
        "/profile": {
            "get": {
                "description": "Returns a user's profile — identity, recent activity visible to the caller (honoring privacy), and their public playlists. Defaults to the caller when username is omitted.",
                "parameters": [
                    {
                        "description": "Subsonic username",
                        "in": "query",
                        "name": "u",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Subsonic password (or t+s token auth)",
                        "in": "query",
                        "name": "p",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Client name",
                        "in": "query",
                        "name": "c",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Target username (defaults to the caller)",
                        "in": "query",
                        "name": "username",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ProfileResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/immerle.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "summary": "User profile",
                "tags": [
                    "users"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
...
}